requirements.txt
absl-py @ file:///home/conda/feedstock_root/build_artifacts/absl-py_1705494584803/work
attrs==24.2.0
Brotli @ file:///C:/b/abs_3d36mno480/croot/brotli-split_1714483178642/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1725278078093/work/certifi
cffi==1.17.1
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
filelock @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/filelock_1701807523603/work
flatbuffers==24.3.25
fonttools==4.54.1
grpcio @ file:///D:/bld/grpc-split_1729252511778/work
idna @ file:///C:/b/abs_aad84bnnw5/croot/idna_1714398896795/work
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1726082825846/work
jax==0.4.35
jaxlib==0.4.35
Jinja2 @ file:///C:/b/abs_92fccttino/croot/jinja2_1716993447201/work
kiwisolver==1.4.7
Markdown @ file:///home/conda/feedstock_root/build_artifacts/markdown_1710435156458/work
MarkupSafe @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/markupsafe_1707425732791/work
matplotlib==3.9.2
mediapipe==0.10.14
ml_dtypes==0.5.0
mpmath @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/mpmath_1699484863771/work
networkx @ file:///C:/b/abs_3bxnu56g9d/croot/networkx_1717597507456/work
numpy @ file:///D:/bld/numpy_1728664377586/work/dist/numpy-2.1.2-cp312-cp312-win_amd64.whl#sha256=7f1422a23098f444b2f2c3e64395641098cb550395f577f537f9e7329847b114
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0
opencv-python-headless==4.10.0
opt_einsum==3.4.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1718189413536/work
pillow @ file:///D:/bld/pillow_1726075241382/work
protobuf==4.25.5
pycparser==2.22
pyparsing==3.2.0
PySocks @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/pysocks_1699473336188/work
python-dateutil==2.9.0.post0
PyYAML @ file:///C:/b/abs_14xkfs39bx/croot/pyyaml_1728657968772/work
requests @ file:///C:/b/abs_9frifg92q2/croot/requests_1721410901096/work
scipy==1.14.1
setuptools==75.1.0
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sounddevice==0.5.1
sympy @ file:///C:/b/abs_4e4p71hdj_/croot/sympy_1724938208509/work
tdqm==0.0.1
tensorboard @ file:///home/conda/feedstock_root/build_artifacts/tensorboard_1727435705122/work/tensorboard-2.18.0-py3-none-any.whl#sha256=107ca4821745f73e2aefa02c50ff70a9b694f39f790b11e6f682f7d326745eab
tensorboard_data_server @ file:///D:/bld/tensorboard-data-server_1728639816333/work/tensorboard_data_server-0.7.0-py3-none-any.whl#sha256=d67bff9fae5ebc80d71e8224dbae4d34256d9c075c51d9090c5a43aa8f927346
torch==2.5.0
torchaudio==2.5.0
torchvision==0.20.0
tqdm==4.66.5
typing_extensions @ file:///C:/b/abs_0as9mdbkfl/croot/typing_extensions_1715268906610/work
urllib3 @ file:///C:/b/abs_9a_f8h_bn2/croot/urllib3_1727769836930/work
Werkzeug @ file:///home/conda/feedstock_root/build_artifacts/werkzeug_1724330738730/work
wheel==0.44.0
win-inet-pton @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/win_inet_pton_1699472992992/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1726248574750/work
